add dataset info (#663)
* Fix import and deprecation issues in unit tests (#871)

* fix some bugs in the unit test of smpl model.

* reorganize `tests/` to solve importing issue (PEP 420)

* fix deprecation warnings in unit tests


* add benchmark regression test script with tmux (#849)

* test the simple case using tmux to run multiple benchmark regression test tasks

* modify and rename the config file and script

* Delete config_list.yaml

* modify the config and rename the filename

* Delete test_benchmark_tmux.py

* modify the script and rename the filename

* Update setup.cfg

* using mmcv.load to avoid introducing the extra dependency on yaml

* fix some typo

* refactor the config file and modify the script accordingly

* modify the config and script

* rename the config file

* Correct dataset preparation guide of WFLW (#873)

* add pr template (#875)

* add CITATION.cff and update setup.py (#876)

* Add copyright header and pre-commit hook (#872)

* Add pre-commit hook to automatically add copyright file header

* update files with copyright header

*  Limit copyright checking in the first 2 lines of a file
* Exclude configs in demo/

* set max-header-lines as 5

* rebase to master and add copyright to new files
* move benchmark_regression into .dev_scripts/benchmark

* Translate tasks/2d_body_keypoint.md (#842)

* 2rd PR remove poseval

* fix lint

* revise the CN version


* fix some bugs in the unit test of smpl model.

* * reorganiz `tests/` to solve importing issue (PEP 420)

* add dataset info

* fix lint

* * fix wrongly modified parts in previous rebase
* fix lint

* rename datasets/_base_ as datasets/base

* resolve compatibility of pose_limb_color

* Add dummy dataset base classes with old names for compatibility

* * Rewrite relative unittest based on dataset_info
* Add bc-breaking test for functions related to dataset_info
* Rename DatasetInfo.dataset_info as DatasetInfo._dataset_info
* Fix dataset_info of h36m dataset

* Handle breaking change pose_limb_color -> pose_link_color

* add unittest for old-fashioned dataset initialization without dataset_info

* resolve naming conflict in unittests

+++ mmpose/datasets/pipelines/bottom_up_transform.py
@@ -118,9 +118,9 @@ def _resize_align_multi_scale_udp(image, input_size, current_scale, min_scale):
 
     trans = get_warp_matrix(
         theta=0,
-        size_input=np.array(scale, dtype=float),
-        size_dst=np.array(size_resized, dtype=float) - 1.0,
-        size_target=np.array(scale, dtype=float))
+        size_input=np.array(scale, dtype=np.float32),
+        size_dst=np.array(size_resized, dtype=np.float32) - 1.0,
+        size_target=np.array(scale, dtype=np.float32))
     image_resized = cv2.warpAffine(
         image.copy(), trans, size_resized, flags=cv2.INTER_LINEAR)
 
@@ -458,8 +458,8 @@ class BottomUpRandomAffine:
                     theta=aug_rot,
                     size_input=center * 2.0,
                     size_dst=np.array(
-                        (_output_size, _output_size), dtype=float) - 1.0,
-                    size_target=np.array((scale, scale), dtype=float))
+                        (_output_size, _output_size), dtype=np.float32) - 1.0,
+                    size_target=np.array((scale, scale), dtype=np.float32))
                 mask[i] = cv2.warpAffine(
                     (mask[i] * 255).astype(np.uint8),
                     trans, (int(_output_size), int(_output_size)),
@@ -473,8 +473,9 @@ class BottomUpRandomAffine:
                 theta=aug_rot,
                 size_input=center * 2.0,
                 size_dst=np.array(
-                    (self.input_size, self.input_size), dtype=float) - 1.0,
-                size_target=np.array((scale, scale), dtype=float))
+                    (self.input_size, self.input_size), dtype=np.float32) -
+                1.0,
+                size_target=np.array((scale, scale), dtype=np.float32))
             image = cv2.warpAffine(
                 image,
                 mat_input, (int(self.input_size), int(self.input_size)),
